requirements.txt
flask
Flask-SQLAlchemy
flask-login
pyrebase4
openai
python-dotenv
boto3